GCP/functions/metrics_collector/requirements.txt
google-cloud-monitoring
google-cloud-storage